Added a horribly coded score controller.

+++ Bumbershoots/Assets/Resources/Scripts/Controllers/CoinController.cs
@@ -4,11 +4,11 @@ using UnityEngine;
 
 public class CoinController : MonoBehaviour {
 
-    public float amountOfCoins;
+    public int amountOfCoins;
 
     void Start(){
-        if(amountOfCoins == 0f){
-            amountOfCoins = 5f;
+        if(amountOfCoins == 0){
+            amountOfCoins = 5;
         }
     }
